InstaGorl/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.4
django-crispy-forms==1.14.0
django-filter==22.1
django-fontawesome==1.0
exceptiongroup==1.1.0
fontawesome==5.10.1.post1
iniconfig==1.1.1
numpy==1.24.1
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.0
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
